[refactor] #1 일기장 필터링과 정렬하는 함수 분리
