+++ src/components/DiaryList.js
@@ -2,17 +2,8 @@ import React, { useState } from "react";
 import MyButton from "./MyButton";
 import { useNavigate } from "react-router-dom";
 import DiaryItem from "./DiaryItem";
-
-const sortOptionList = [
-  { value: "latest", name: "최신순" },
-  { value: "oldest", name: "오래된순" },
-];
-
-const filterOptionList = [
-  { value: "all", name: "전부 다" },
-  { value: "good", name: "좋은 감정만" },
-  { value: "bad", name: "안좋은 감정만" },
-];
+import { sortOptionList, filterOptionList } from "../const/options";
+import { getProcessedDiaryList } from "../utils/filtersortList";
 
 const ControlMenu = React.memo(({ value, onChange, optionList }) => {
   return (
@@ -36,40 +27,7 @@ const DiaryList = ({ diaryList }) => {
   const [sortType, setSortType] = useState("latest");
   const [filter, setFilter] = useState("all");
 
-  const getProcessedDiaryList = () => {
-    const filterCallback = (item) => {
-      if (filter === "good") {
-        return parseInt(item.emotion) <= 3;
-      } else {
-        return parseInt(item.emotion) > 3;
-      }
-    };
-
-    const compare = (a, b) => {
-      const dateSortBA = parseInt(b.date) - parseInt(a.date);
-      const dateSortAB = parseInt(a.date) - parseInt(b.date);
-      if (sortType === "latest") {
-        return dateSortBA === 0
-          ? parseInt(b.diaryId) - parseInt(a.diaryId)
-          : dateSortBA;
-      } else {
-        return dateSortAB === 0
-          ? parseInt(a.diaryId) - parseInt(b.diaryId)
-          : dateSortAB;
-      }
-    };
-
-    const copyList = JSON.parse(JSON.stringify(diaryList));
-
-    //오래된순, 최신순으로 정렬된 데이터를 또 필터링 해주기
-    //1,2,3 => good, | 4, 5 => bad
-
-    const filteredList =
-      filter === "all" ? copyList : copyList.filter((it) => filterCallback(it));
-
-    const sortedList = filteredList.sort(compare);
-    return sortedList;
-  };
+  const processedDiaryList = getProcessedDiaryList(diaryList, filter, sortType);
 
   return (
     <div className="DiaryList">
@@ -95,7 +53,7 @@ const DiaryList = ({ diaryList }) => {
         </div>
       </div>
 
-      {getProcessedDiaryList().map((it) => (
+      {processedDiaryList.map((it) => (
         <DiaryItem key={it.diaryId} {...it} />
       ))}
     </div>
